fix: dangerous account commands

+++ src/CoreLib/Game/Commands/Protocols/CommandAccountProtocol.cs
@@ -42,6 +42,12 @@ internal class CommandAccountProtocol : CommandProtocol {
             return;
         }
 
+        // Yeah, you can't delete your own account.
+        if (username == Context.Account.Username) {
+            InformSenderClient("You cannot delete your own account.");
+            return;
+        }
+
         AccountCollection.DeleteAccount(username);
         InformSenderClient("Account deleted successfully.");
     }
@@ -108,6 +114,19 @@ internal class CommandAccountProtocol : CommandProtocol {
         var account = AccountCollection.GetAccount(username);
         if (account is null) {
             InformSenderClient("Account not found.");
+            return;
+        }
+
+        // You cannot change the auth level of an account with a higher auth level than you.
+        var authorityReason = $"{Context.Account.Username} wants to change the auth level of account {username}.";
+        if (!AuthorityRequester.RequestAuthority(account.AuthLevel, Context.Account, authorityReason)) {
+            InformSenderClient("You cannot change the auth level of an account with a higher auth level than you.");
+            return;
+        }
+
+        // You cannot change the auth level of your own account.
+        if (username == Context.Account.Username) {
+            InformSenderClient("You cannot change the auth level of your own account.");
             return;
         }
